backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
google-generativeai>=0.3
django-ratelimit>=4.1
bleach>=6.1
gunicorn>=21.0
whitenoise>=6.6
requests>=2.31
